fix float

+++ ai_user/prompts/image_prompt/base.py
@@ -47,5 +47,5 @@ class BaseImagePrompt(Prompt):
         image_resolution = width * height
         if image_resolution > target_resolution:
             scale_factor = (target_resolution / image_resolution) ** 0.5
-            return image.resize((width * scale_factor, height * scale_factor), Image.Resampling.LANCZOS)
+            return image.resize((int(width * scale_factor), int(height * scale_factor)), Image.Resampling.LANCZOS)
         return image
